eliminar registros de cursos y pagos

+++ src/main/java/com/sae/gandhi/spring/service/impl/CursoCostosServiceImpl.java
@@ -1,6 +1,7 @@
 package com.sae.gandhi.spring.service.impl;
 
 import java.util.Calendar;
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
@@ -8,8 +9,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.sae.gandhi.spring.dao.AlumnoCursoDAO;
+import com.sae.gandhi.spring.dao.AlumnoPagoDAO;
 import com.sae.gandhi.spring.dao.CursoCostosDAO;
+import com.sae.gandhi.spring.dao.CursosDAO;
+import com.sae.gandhi.spring.entity.AlumnoCurso;
+import com.sae.gandhi.spring.entity.AlumnoPagos;
 import com.sae.gandhi.spring.entity.CursoCostos;
+import com.sae.gandhi.spring.entity.Cursos;
+import com.sae.gandhi.spring.service.AlumnoCursoService;
 import com.sae.gandhi.spring.service.CursoCostosService;
 import com.sae.gandhi.spring.utils.builder.CursoCostoBuilder;
 import com.sae.gandhi.spring.vo.CursoCostosVO;
@@ -20,6 +28,18 @@ public class CursoCostosServiceImpl implements CursoCostosService {
 	
 	@Autowired
 	private CursoCostosDAO cursoCostosDAO;
+	
+	@Autowired
+	private AlumnoPagoDAO alumnoPagoDAO;
+	
+	@Autowired
+	private AlumnoCursoDAO alumnoCursoDAO;
+	
+	@Autowired
+	private CursosDAO cursosDAO;
+	
+	@Autowired
+	private AlumnoCursoService alumnoCursoService;
 
 	@Override
 	public List<CursoCostosVO> findAllActive() {
@@ -40,14 +60,30 @@ public class CursoCostosServiceImpl implements CursoCostosService {
 	@Override
 	public void save(CursoCostosVO cursoCostosVo) {
 		CursoCostos cursoCostos = CursoCostoBuilder.createCursoCosto(cursoCostosVo);
+		Calendar cal = Calendar.getInstance();
+		
 		if(cursoCostosVo.getCostosVO()!=null){
 			cursoCostos.setCostoId(cursoCostosVo.getCostosVO().getCostoId());			
 		}else{
 			cursoCostos.setCostoId(cursoCostosVo.getCostoId());
 		}
 		cursoCostos.setCursoCostoActivo(true);
-		cursoCostos.setFechaCreacion(Calendar.getInstance().getTime());
-		cursoCostosDAO.save(cursoCostos);
+		cursoCostos.setFechaCreacion(cal.getTime());
+		cursoCostos = cursoCostosDAO.save(cursoCostos);
+		
+		//Guardar el curso costo en los alumnos
+		Optional<Cursos> optional = cursosDAO.findById(cursoCostos.getCursoId());
+		List<AlumnoCurso> lst = alumnoCursoDAO.findByCursoId(cursoCostos.getCursoId());
+		Cursos curso = optional.orElse(null);
+		Date initDate = curso.getCursoFechaInicio();
+		
+		if(cal.getTime().after(curso.getCursoFechaInicio())){
+			initDate = cal.getTime(); 
+		}
+		
+		for(AlumnoCurso alumnoCurso : lst){
+			alumnoCursoService.createStudentPayment(cursoCostos, alumnoCurso, initDate, curso.getCursoFechaFin());
+		}
 	}
 
 	@Override
@@ -67,8 +103,18 @@ public class CursoCostosServiceImpl implements CursoCostosService {
 	}
 
 	@Override
-	public void delete(Integer cursoCostoId) {
+	public boolean delete(Integer cursoCostoId) {
+		
+		List<AlumnoPagos> lst = alumnoPagoDAO.findByCursoCostoIdPayed(cursoCostoId);
+		//Busca si algún alumno ha pagado el costo del curso en algún mes
+		if(lst!=null && !lst.isEmpty()){
+			return false;
+		}
+		
+		//elimina todos los pagos de los alumnos que sean del cursoCostoId
+		alumnoPagoDAO.deleteByCursoCostoId(cursoCostoId);
 		cursoCostosDAO.deleteById(cursoCostoId);
+		return true;
 	}
 
 	@Override
